requirements.txt
decorator==4.0.6
Django==1.9
django-filter==0.11.0
djangorestframework==3.3.1
praw==3.3.0
psycopg2==2.6.1
requests==2.8.1
six==1.10.0
update-checker==0.11
virtualenv==13.1.2
wheel==0.26.0
